+++ lib/features/aicycle_buy_me/data/remote_data/aicycle_buyme_api.dart
@@ -36,7 +36,7 @@ class AicycleBuyMeApi extends APIRequest {
             endpoint: Endpoint.claimFolders,
             method: HTTPMethod.get,
             isLogResponse: true,
-            body: {
+            query: {
               "externalClaimId": externalClaimId,
             });
 }