+++ lib/features/aicycle_buy_me/presentation/aicycle_buy_me.dart
@@ -5,23 +5,38 @@ import 'package:get/get.dart';
 import '../../../aicycle_buyme_lib.dart';
 import '../../../enum/app_state.dart';
 import '../../../generated/assets.gen.dart';
-import '../../common/app_string.dart';
+import '../../../generated/locales.g.dart';
 import '../../common/base_widget.dart';
 import '../../common/themes/c_colors.dart';
 import '../../common/themes/c_textstyle.dart';
 import '../../folder_details/data/models/buy_me_image_model.dart';
 import '../../folder_details/presentation/folder_detail_page.dart';
+import '../../common/extension/translation_ext.dart';
 import 'aicycle_buy_me_controller.dart';
 
+enum Evn {
+  dev,
+  stage,
+  production,
+}
+
 String? apiToken;
+Evn environtment = Evn.production;
+Locale? locale;
 
 class AiCycleBuyMeArgument {
   final String externalClaimId;
   final String apiToken;
+  final Evn? environtment;
+  final Locale? locale;
+  final String? aicycleClaimId;
 
   AiCycleBuyMeArgument({
     required this.externalClaimId,
     required this.apiToken,
+    this.environtment,
+    this.locale,
+    this.aicycleClaimId,
   });
 }
 
@@ -53,14 +68,16 @@ class _AiCycleBuyMeState
     super.initState();
     controller.argument = widget.argument;
     apiToken = widget.argument.apiToken;
+    environtment = widget.argument.environtment ?? Evn.production;
+    locale = widget.argument.locale;
     controller.status.listen((state) {
       if (state.state == AppState.redirect) {
         Navigator.of(context).pushReplacement(
           MaterialPageRoute<void>(
             builder: (BuildContext context) => FolderDetailPage(
-              claimFolderId: controller.claimFolder.value?.claimId ?? '',
-              externalClaimId: widget.argument.externalClaimId,
+              argument: controller.argument,
               onViewResultCallBack: widget.onViewResultCallBack,
+              onCallEngineSuccessfully: (p0) {},
             ),
           ),
         );
@@ -70,6 +87,7 @@ class _AiCycleBuyMeState
 
   @override
   Widget build(BuildContext context) {
+    super.build(context);
     return Scaffold(
       body: Container(
         height: double.infinity,
@@ -94,7 +112,7 @@ class _AiCycleBuyMeState
             ),
             const Gap(24),
             Text(
-              AppString.pleaseWait,
+              LocaleKeys.pleaseWait.trans,
               style: CTextStyles.base.s16.whiteColor,
             ),
             const Gap(12),
@@ -110,4 +128,7 @@ class _AiCycleBuyMeState
       ),
     );
   }
+
+  @override
+  bool get wantKeepAlive => false;
 }