+++ lib/features/aicycle_buy_me/presentation/aicycle_buy_me_controller.dart
@@ -36,7 +36,7 @@ class AiCycleBuyMeController extends BaseController {
     res.fold(
       (l) {
         if (l.details.toString().toLowerCase().contains('duplicate')) {
-          getDupliacateFolder();
+          getDuplicateFolder();
         } else {
           isLoading(false);
           status.value = BaseStatus(
@@ -46,6 +46,13 @@ class AiCycleBuyMeController extends BaseController {
         }
       },
       (r) {
+        argument = AiCycleBuyMeArgument(
+          externalClaimId: argument.externalClaimId,
+          apiToken: argument.apiToken,
+          aicycleClaimId: r.claimId,
+          environtment: argument.environtment,
+          locale: argument.locale,
+        );
         isLoading(false);
         status.value = BaseStatus(
           message: null,
@@ -56,7 +63,7 @@ class AiCycleBuyMeController extends BaseController {
     );
   }
 
-  void getDupliacateFolder() async {
+  void getDuplicateFolder() async {
     var res = await getDuplicateFolderUsecase(
         externalClaimId: argument.externalClaimId);
     res.fold(
@@ -68,6 +75,13 @@ class AiCycleBuyMeController extends BaseController {
         );
       },
       (r) {
+        argument = AiCycleBuyMeArgument(
+          externalClaimId: argument.externalClaimId,
+          apiToken: argument.apiToken,
+          aicycleClaimId: r.claimId,
+          environtment: argument.environtment,
+          locale: argument.locale,
+        );
         isLoading(false);
         status.value = BaseStatus(
           message: null,