+++ lib/features/common/c_loading_view.dart
@@ -1,5 +1,7 @@
 import 'dart:io';
 
+import 'package:aicycle_buyme_lib/features/common/extension/translation_ext.dart';
+import 'package:aicycle_buyme_lib/generated/locales.g.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
@@ -59,7 +61,7 @@ class LoadingView<T> extends StatelessWidget {
               : RotatedBox(
                   quarterTurns: quarterTurns,
                   child: Container(
-                    height: 154,
+                    height: 180,
                     width: quarterTurns == 0 || quarterTurns == 2
                         ? double.maxFinite
                         : 300,
@@ -89,7 +91,7 @@ class LoadingView<T> extends StatelessWidget {
                         ),
                         const Gap(8),
                         Text(
-                          'Đang xử lý dữ liệu, vui lòng chờ.',
+                          LocaleKeys.isProcessing.trans,
                           style: CTextStyles.base.s14.w500(),
                           textAlign: TextAlign.center,
                         )