+++ CHANGELOG.md
@@ -1,3 +1,68 @@
 ## 0.0.1
 
-* First version 
+* First version
+
+## 0.0.2
+
+* Add API Key
+
+## 0.0.3
+
+* Fix bugs
+
+## 0.0.4
+
+* Fix package name
+
+## 0.0.5
+
+* Fix camera loading bug
+
+## 0.0.6
+
+* Fix environment
+
+## 0.0.7
+
+* rotate capture image
+
+## 0.0.8
+
+* Support vi_VN, en_US, ja_JP languages
+* Fix bugs
+
+## 0.0.9
+
+* Fix bugs
+
+## 0.1.0
+
+* Fix multiple languages
+
+## 0.1.1
+
+* Custom for Aicycle
+
+## 0.1.2
+
+* Hotfix UI bug
+
+## 0.1.3
+
+* Hotfix camera UI
+
+## 0.1.4
+
+* Update rotate image
+
+## 0.1.5
+
+* Fix bugs
+
+## 0.1.6
+
+* Fix bugs
+
+## 0.1.7
+
+* Bug fixes and improvements